requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.5
drf-yasg==1.21.7
prometheus-client==0.19.0
structlog==23.2.0
psutil==7.0.0
requests==2.31.0
dotenv==0.9.9